* main: (52 commits)
  ENH: stats.vonmises.fit: treat case in which location likelihood equation has no solution(#18190)
  MAINT: stats.kendalltau: avoid overflow (#18193)
  DOC: Optimize: Fix for side bar rendering on top of Hessian (#18189)
  MAINT: optimize.linprog: fix bound checks for integrality > 1 (#18160)
  MAINT: Windows distutils cdist/pdist shims (#18169)
  BUG: interpolate: add x-y length validation for `make_smoothing_spline`. (#18188)
  ENH: Added `_sf` method for anglit distribution (#17832) (#18178)
  DOC: Fixed missing curly bracket in scipy.css
  DOC: Improving wording and docs for legacy directive
  DOC: Move legacy directive to not be first in the file
  DOC: Legacy directive custom styling
  DOC: Add optional argument to Legacy directive
  DOC: Ignore legacy directive in refguide_check
  DOC: Documenting the usage of the legacy directive
  DOC: Add legacy directive for documentation
  MAINT: stats.ecdf: store number at risk just before events (#18187)
  DOC: cite pip issue about multiple `--config-settings` (#18174)
  DOC: update links for ARPACK to point to ARPACK-NG (#18173)
  MAINT: stats.logistic.fit: simplify
  MAINT: optimize.root_scalar: return gracefully when callable returns NaN (#18172)
  ...

+++ scipy/stats/_stats_py.py
@@ -5821,9 +5821,10 @@ def kendalltau(x, y, initial_lexsort=None, nan_policy='propagate',
     def count_rank_tie(ranks):
         cnt = np.bincount(ranks).astype('int64', copy=False)
         cnt = cnt[cnt > 1]
-        return ((cnt * (cnt - 1) // 2).sum(),
-                (cnt * (cnt - 1.) * (cnt - 2)).sum(),
-                (cnt * (cnt - 1.) * (2*cnt + 5)).sum())
+        # Python ints to avoid overflow down the line
+        return (int((cnt * (cnt - 1) // 2).sum()),
+                int((cnt * (cnt - 1.) * (cnt - 2)).sum()),
+                int((cnt * (cnt - 1.) * (2*cnt + 5)).sum()))
 
     size = x.size
     perm = np.argsort(y)  # sort on y and convert y to dense ranks
@@ -5840,7 +5841,7 @@ def kendalltau(x, y, initial_lexsort=None, nan_policy='propagate',
     obs = np.r_[True, (x[1:] != x[:-1]) | (y[1:] != y[:-1]), True]
     cnt = np.diff(np.nonzero(obs)[0]).astype('int64', copy=False)
 
-    ntie = (cnt * (cnt - 1) // 2).sum()  # joint ties
+    ntie = int((cnt * (cnt - 1) // 2).sum())  # joint ties
     xtie, x0, x1 = count_rank_tie(x)     # ties in x, stats
     ytie, y0, y1 = count_rank_tie(y)     # ties in y, stats
